[flaky-ci] Run cancellation rewind tests in NativeAOT
Move the cancellation rewind regression and its focused helpers out of the SSL-tagged handler fixture so PublishAot lanes execute them.
